+++ rules/early-return/tests/Rector/If_/ChangeOrIfContinueToMultiContinueRector/Fixture/fixture.php.inc
@@ -1,13 +1,13 @@
 <?php
 
-namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeAndIfContinueToMultiContinueRector\Fixture;
+namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeOrIfContinueToMultiContinueRector\Fixture;
 
 class Fixture
 {
     public function canDrive(Car $newCar)
     {
         foreach ($cars as $car) {
-            if ($car->hasWheels() && $car->hasFuel()) {
+            if ($car->hasWheels() || $car->hasFuel()) {
                 continue;
             }
             $car->setWheel($newCar->wheel);
@@ -20,17 +20,17 @@ class Fixture
 -----
 <?php
 
-namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeAndIfContinueToMultiContinueRector\Fixture;
+namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeOrIfContinueToMultiContinueRector\Fixture;
 
 class Fixture
 {
     public function canDrive(Car $newCar)
     {
         foreach ($cars as $car) {
-            if (!$car->hasWheels()) {
+            if ($car->hasWheels()) {
                 continue;
             }
-            if (!$car->hasFuel()) {
+            if ($car->hasFuel()) {
                 continue;
             }
             $car->setWheel($newCar->wheel);
