+++ rules/early-return/tests/Rector/If_/ChangeOrIfContinueToMultiContinueRector/Fixture/skip_and.php.inc
@@ -1,14 +1,13 @@
 <?php
 
-namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeAndIfContinueToMultiContinueRector\Fixture;
+namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeOrIfContinueToMultiContinueRector\Fixture;
 
-class SkipMultiStmts
+class SkipAnd
 {
     public function canDrive(Car $newCar)
     {
         foreach ($cars as $car) {
             if ($car->hasWheels() && $car->hasFuel()) {
-                executeSideEffect();
                 continue;
             }
