+++ rules/early-return/tests/Rector/If_/ChangeOrIfContinueToMultiContinueRector/Fixture/skip_multi_stmts.php.inc
@@ -1,13 +1,14 @@
 <?php
 
-namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeAndIfContinueToMultiContinueRector\Fixture;
+namespace Rector\EarlyReturn\Tests\Rector\If_\ChangeOrIfContinueToMultiContinueRector\Fixture;
 
-class SkipNotAnd
+class SkipMultiStmts
 {
     public function canDrive(Car $newCar)
     {
         foreach ($cars as $car) {
             if ($car->hasWheels() || $car->hasFuel()) {
+                executeSideEffect();
                 continue;
             }
